fix: 타입 충돌 해결

+++ src/main/java/com/example/delivery/domain/order/dto/Response/Cart/CartItemDetailResponseDto.java
@@ -8,9 +8,9 @@ public class CartItemDetailResponseDto {
     Long menuId;
     String menuName;
     String intro;
-    Long price;
-    int quantity;
-    Long totalPrice;
+    Integer price;
+    Integer quantity;
+    Integer totalPrice;
 
     public CartItemDetailResponseDto(CartItem item) {
         this.menuId = item.getMenu().getId();
